Fix: Resolve module and type errors
The commit addresses two TypeScript errors: the first indicates that the `Index` module in `src/App.tsx` is not exporting a default, and the second indicates that the `src/pages/Index.tsx` file cannot find the `./types` module. The fix will involve ensuring a default export for the `Index` module and resolving the import of the missing types module.
